+++ lib/chef/provider.rb
@@ -22,8 +22,8 @@ require "chef/mixin/convert_to_class_name"
 require "chef/mixin/enforce_ownership_and_permissions"
 require "chef/mixin/why_run"
 require "chef/mixin/shell_out"
-require "chef/mixin/powershell_out"
 require "chef/mixin/provides"
+require "chef/dsl/core"
 require "chef/platform/service_helpers"
 require "chef/node_map"
 require "forwardable"
@@ -31,12 +31,12 @@ require "forwardable"
 class Chef
   class Provider
     require "chef/mixin/why_run"
-    require "chef/mixin/shell_out"
     require "chef/mixin/provides"
     include Chef::Mixin::WhyRun
-    include Chef::Mixin::ShellOut
-    include Chef::Mixin::PowershellOut
     extend Chef::Mixin::Provides
+
+    # includes the "core" DSL and not the "recipe" DSL by design
+    include Chef::DSL::Core
 
     # supports the given resource and action (late binding)
     def self.supports?(resource, action)
